+++ Runtime/DependencyContext.cs
@@ -1,18 +1,26 @@
 ﻿namespace AncientLightStudios.Nanoject
 {
     using System;
+    using System.Collections;
     using System.Collections.Generic;
     using System.Linq;
     using System.Reflection;
-
+    using System.Text;
+    using JetBrains.Annotations;
+    using UnityEngine;
 
     public sealed class DependencyContext
     {
-        private readonly Dictionary<Type, Dictionary<string, object>> _dependencies =
-            new Dictionary<Type, Dictionary<string, object>>();
+        private readonly HolderByTypeAndQualifier<object> _resolvedComponents =
+            new HolderByTypeAndQualifier<object>();
 
-        private readonly Dictionary<Type, Dictionary<string, object>> _unresolvedDependencies =
-            new Dictionary<Type, Dictionary<string, object>>();
+        private readonly HolderByTypeAndQualifier<DependencySet> _collections =
+            new HolderByTypeAndQualifier<DependencySet>();
+
+        private readonly HolderByTypeAndQualifier<UnresolvedComponent> _unresolvedComponents =
+            new HolderByTypeAndQualifier<UnresolvedComponent>();
+
+        public bool IsResolved => _unresolvedComponents.Count == 0;
 
         public DependencyContext()
         {
@@ -40,193 +48,453 @@
             DeclareInternal(type, instance);
         }
 
-        
+
         private void DeclareInternal(Type type, object instance = default, string qualifier = "")
         {
-            if (!_unresolvedDependencies.TryGetValue(type, out var values))
-            {
-                values = new Dictionary<string, object>();
-                _unresolvedDependencies[type] = values;
-            }
-
-            values[qualifier] = instance;
+            var unresolvedDependency =
+                instance == default ? new UnresolvedComponent(type) : new UnresolvedComponent(instance);
+            _unresolvedComponents.Put(type, qualifier, unresolvedDependency);
         }
-
-		public bool IsDeclared(Type type, string qualifier = "") {
-			if (_unresolvedDependencies.TryGetValue(type, out var values)) {
-				if (values.ContainsKey(qualifier)) {
-					return true;
-				}
-			}
-			if (_dependencies.TryGetValue(type, out var dependenciesByQualifier)) {
-				if (dependenciesByQualifier.ContainsKey(qualifier)) {
-					return true;
-				}
-			}
-			return false;
-		}
-
 
         public T Get<T>(string qualifier = "")
         {
-            if (_unresolvedDependencies.Count > 0)
+            if (!IsResolved)
             {
                 throw new InvalidOperationException("This dependency context is not resolved.");
             }
 
-            var type = typeof(T);
-            if (_dependencies.TryGetValue(type, out var dependenciesByQualifier))
+            if (_resolvedComponents.TryGetUniqueValue(typeof(T), qualifier, out var result))
             {
-                if (dependenciesByQualifier.TryGetValue(qualifier, out var result))
-                {
-                    return (T) result;
-                }
+                return (T) result;
             }
 
-            throw new InvalidOperationException($"No dependency of type {type} with qualifier '{qualifier}'.");
+            throw new InvalidOperationException(
+                $"No unique dependency of type {typeof(T)} with qualifier '{qualifier}'.");
         }
 
         public List<T> GetAll<T>()
         {
-            var result = new List<T>();
-            var resultType = typeof(T);
-
-            var types = new List<Type>(_dependencies.Keys);
-            foreach (var type in types)
-            {
-                if (resultType.IsAssignableFrom(type))
-                {
-                    foreach (var value in _dependencies[type].Values)
-                    {
-                        result.Add((T) value);
-                    }
-                }
-            }
-
-            return result;
+            return _resolvedComponents.Where((type, qualifier) => typeof(T).IsAssignableFrom(type)).Cast<T>().ToList();
         }
+
 
         public void Resolve()
         {
-            bool hasProgress;
+            bool ResolveUnique(Type type, string qualifier, out object result)
+            {
+                if (_resolvedComponents
+                    .Where((itsType, itsQualifier) => type.IsAssignableFrom(itsType) && itsQualifier == qualifier)
+                    .UniqueMatch(out result))
+                {
+                    return true;
+                }
 
+                if (Utils.IsCollectionType(type, out var collectionContentType))
+                {
+                    // first check if there are any unresolved components of the given collection content type
+                    // or an assignable type. we need to do this to make sure we don't inject a half-finished collection
+                    // into a newly created dependency. We can only create collections after all components that
+                    // make up the collection are resolved.
+                    if (!_unresolvedComponents.Where((itsType, _) => collectionContentType.IsAssignableFrom(itsType))
+                        .Any())
+                    {
+                        // if we already built a collection for this type and qualifier, use this.
+                        if (!_collections.Where((itsType, itsQualifier) =>
+                            collectionContentType == itsType && qualifier == itsQualifier).UniqueMatch(out var set))
+                        {
+                            // otherwise build a new one
+                            set = Utils.MakeDependencySet(collectionContentType);
+                            _collections.Put(type, qualifier, set);
+                        }
+
+                        // make sure the set contains all known matching components. Since it is a set, it doesn't
+                        // matter if we add a component more than once.
+                        foreach (var item in _resolvedComponents.Where((itsType, itsQualifier) =>
+                            collectionContentType.IsAssignableFrom(itsType)
+                            && (qualifier == "" || qualifier == itsQualifier))
+                        )
+                        {
+                            set.Add(item);
+                        }
+
+                        result = set;
+                        return true;
+                    }
+                }
+
+                result = default;
+                return false;
+            }
+
+            bool hasProgress;
             do
             {
                 hasProgress = false;
 
-                var unresolvedTypes = new List<Type>(_unresolvedDependencies.Keys);
-                foreach (var unresolvedType in unresolvedTypes)
+                // try a round of resolving stuff.
+                _unresolvedComponents.ForEach((unresolvedType, unresolvedQualifier, unresolvedComponent) =>
                 {
-                    var unresolvedByQualifier = _unresolvedDependencies[unresolvedType];
-                    var unresolvedQualifiers = new List<string>(unresolvedByQualifier.Keys);
-
-                    foreach (var unresolvedQualifier in unresolvedQualifiers)
+                    if (!unresolvedComponent.TryResolve(ResolveUnique, out var instance))
                     {
-                        // if this object has been given in advance, so we just use it
-                        var instance = unresolvedByQualifier[unresolvedQualifier] ?? TryConstruct(unresolvedType, unresolvedQualifier);
-
-                        if (instance == default)
-                        {
-                            continue; // we cannot resolve it right now, so skip it for next round
-                        }
-
-                        // try to resolve late init method, if it exists
-                        instance = TryResolveLateInitializer(unresolvedType, instance);
-                        
-                        if (instance == default)
-                        {
-                            continue; // we cannot resolve it right now, so skip it for next round
-                        }
-                        
-                        // we have resolved it, so we can add it to the known dependencies
-                        if (!_dependencies.TryGetValue(unresolvedType, out var values))
-                        {
-                            values = new Dictionary<string, object>();
-                            _dependencies[unresolvedType] = values;
-                        }
-                        values[unresolvedQualifier] = instance;
-
-                        hasProgress = true;
-
-                        // remove it from the nested qualifiers dictionary
-                        unresolvedByQualifier.Remove(unresolvedQualifier);
-
-                        // and remove the whole qualifiers dictionary if it is empty.
-                        if (unresolvedByQualifier.Count == 0)
-                        {
-                            _unresolvedDependencies.Remove(unresolvedType);
-                        }
+                        return;
                     }
-                }
+
+                    // we have resolved it, so we can add it to the known dependencies
+                    _resolvedComponents.Put(unresolvedType, unresolvedQualifier, instance);
+
+                    // and remove it from the unresolved dependencies
+                    _unresolvedComponents.Remove(unresolvedType, unresolvedQualifier, unresolvedComponent);
+                    hasProgress = true;
+                });
             } while (hasProgress);
 
-            if (_unresolvedDependencies.Count <= 0)
+            if (_unresolvedComponents.Count <= 0)
             {
                 return;
             }
-            
-            var unresolved = "";
-            foreach (var keyValuePair in _unresolvedDependencies)
-            {
-                unresolved +=  $"Unresolved dependency of type {keyValuePair.Key.FullName}\n";
-            }
 
-            throw new InvalidOperationException($"There are unresolved dependencies left!: \n{unresolved}");
+            _unresolvedComponents.ForEach((itsType, itsQualifier, item) =>
+            {
+                var actualQualifier = itsQualifier == "" ? "<no qualifier>" : itsQualifier;
+                Debug.Log($"[{actualQualifier}] {itsType.FullName}\n{item}");
+            });
+
+            throw new InvalidOperationException("There are unresolved dependencies left!");
+        }
+    }
+
+    internal abstract class DependencySet
+    {
+        public abstract void Add(object toAdd);
+    }
+
+    internal class TypedDependencySet<T> : DependencySet, IReadOnlyCollection<T>
+    {
+        private readonly HashSet<T> _set = new HashSet<T>();
+
+        public override void Add(object toAdd)
+        {
+            _set.Add((T) toAdd);
         }
 
-        private object TryResolveLateInitializer(Type type, object instance)
+        IEnumerator IEnumerable.GetEnumerator()
         {
-            var methods = type.GetMethods(BindingFlags.Public | BindingFlags.Instance)
+            return _set.GetEnumerator();
+        }
+
+        public IEnumerator<T> GetEnumerator()
+        {
+            return _set.GetEnumerator();
+        }
+
+        public int Count => _set.Count;
+    }
+
+    /// <summary>
+    /// Various helper functions that don't fit anywhere else.
+    /// </summary>
+    internal static class Utils
+    {
+        /// <summary>
+        /// The supported collection type.
+        /// </summary>
+        private static readonly Type CollectionType = typeof(IReadOnlyCollection<>);
+
+        /// <summary>
+        /// An empty list, so we don't need to create unnecessary instances.
+        /// </summary>
+        public static readonly ICollection<string> EmptyList = new string[0];
+
+        /// <summary>
+        /// Checks if the given type is a supported collection type for injection. Supported is any <see cref="IReadOnlyCollection{T}"/>.
+        /// </summary>
+        /// <param name="type">the type argument</param>
+        /// <param name="collectionContentType">the content type of the collection (e.g. what is in the collection).</param>
+        /// <returns></returns>
+        public static bool IsCollectionType(Type type, out Type collectionContentType)
+        {
+            if (type.IsGenericType && type.GetGenericTypeDefinition() == CollectionType)
+            {
+                collectionContentType = type.GetGenericArguments()[0];
+                return true;
+            }
+
+            collectionContentType = default;
+            return false;
+        }
+
+        /// <summary>
+        /// Create a set to hold dependencies of the given type.
+        /// </summary>
+        /// <param name="t">The type that the set should hold.</param>
+        /// <returns>a DependencySet that can hold dependencies of the given type.</returns>
+        public static DependencySet MakeDependencySet(Type t)
+        {
+            return (DependencySet) Activator.CreateInstance(typeof(TypedDependencySet<>).MakeGenericType(t));
+        }
+
+        /// <summary>
+        /// Returns true if the given enumerable only contains one element, false otherwise. Returns the single element
+        /// inside the enumerable.
+        /// </summary>
+        public static bool UniqueMatch<T>(this IEnumerable<T> input, out T result)
+        {
+            var list = input.ToList();
+            if (list.Count != 1)
+            {
+                result = default;
+                return false;
+            }
+
+            result = list[0];
+            return true;
+        }
+    }
+
+
+    /// <summary>
+    /// Helper class for our nested dictionary structures to avoid some typing.
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
+    internal class HolderByTypeAndQualifier<T>
+    {
+        private readonly Dictionary<Type, Dictionary<string, HashSet<T>>> _value =
+            new Dictionary<Type, Dictionary<string, HashSet<T>>>();
+
+        public int Count => _value.Count;
+
+        public IEnumerable<T> Where(Func<Type, string, bool> filter)
+        {
+            return _value.SelectMany(outer => outer.Value, (outer, inner) => new {outer, inner})
+                .Where(t => filter(t.outer.Key, t.inner.Key))
+                .SelectMany(t => t.inner.Value);
+        }
+
+        public void ForEach(Action<Type, string, T> handler)
+        {
+            foreach (var outer in _value.ToList())
+            {
+                foreach (var inner in outer.Value.ToList())
+                {
+                    foreach (var item in inner.Value.ToList())
+                    {
+                        handler(outer.Key, inner.Key, item);
+                    }
+                }
+            }
+        }
+
+        public void Put(Type type, string qualifier, T value)
+        {
+            if (!_value.TryGetValue(type, out var byQualifier))
+            {
+                byQualifier = new Dictionary<string, HashSet<T>>();
+                _value[type] = byQualifier;
+            }
+
+            if (!byQualifier.TryGetValue(qualifier, out var set))
+            {
+                set = new HashSet<T>();
+                byQualifier[qualifier] = set;
+            }
+
+            set.Add(value);
+        }
+
+        public bool TryGetValue(Type type, string qualifier, out HashSet<T> result)
+        {
+            if (_value.TryGetValue(type, out var byQualifier))
+            {
+                return byQualifier.TryGetValue(qualifier, out result);
+            }
+
+            result = default;
+            return false;
+        }
+
+        public bool TryGetUniqueValue(Type type, string qualifier, out T result)
+        {
+            if (TryGetValue(type, qualifier, out var set))
+            {
+                if (set.Count == 1)
+                {
+                    result = set.First();
+                    return true;
+                }
+            }
+
+            result = default;
+            return false;
+        }
+
+        public void Remove(Type type, string qualifier, T value)
+        {
+            if (!_value.TryGetValue(type, out var byQualifier))
+            {
+                return;
+            }
+
+            if (!byQualifier.TryGetValue(qualifier, out var set))
+            {
+                return;
+            }
+
+            set.Remove(value);
+
+            if (set.Count == 0)
+            {
+                byQualifier.Remove(qualifier);
+            }
+
+            if (byQualifier.Count == 0)
+            {
+                _value.Remove(type);
+            }
+        }
+    }
+
+    /// <summary>
+    /// Class representing a component that still needs to be resolved. Using this class allows to cache introspection
+    /// results and do a gradual resolve while saving intermediate resolution steps. This makes the whole process faster
+    /// and also allows for better debugging.
+    /// </summary>
+    internal class UnresolvedComponent
+    {
+        public delegate bool UniqueResolver(Type type, string qualifier, out object result);
+
+        /// <summary>
+        /// The existing instance of the dependency. Only set for objects with [LateInit] initializers.
+        /// </summary>
+        [CanBeNull] private readonly object _existingInstance;
+
+        /// <summary>
+        /// The initializer that should be used to initialize the instance (either a constructor or a [LateInit] method).
+        /// For dependencies with existing instances this can be null if the dependency has no [LateInit] method.
+        /// </summary>
+        [CanBeNull] private readonly MethodBase _initializer;
+
+        /// <summary>
+        /// The types of all parameters to be resolved.
+        /// </summary>
+        private readonly Type[] _types;
+
+        /// <summary>
+        /// The qualifiers of all parameters to be resolved.
+        /// </summary>
+        private readonly string[] _qualifiers;
+
+        /// <summary>
+        /// All currently resolved parameters.
+        /// </summary>
+        private readonly object[] _resolvedDependencies;
+
+        /// <summary>
+        /// Flags indicating if a certain parameter is already resolved.
+        /// </summary>
+        private readonly bool[] _isResolved;
+
+        /// <summary>
+        /// Safety flag to avoid resolving this component more than once.
+        /// </summary>
+        private bool _wasResolved;
+
+
+        /// <summary>
+        /// Constructor for [LateInit] components.
+        /// </summary>
+        /// <param name="existingInstance">the existing instance</param>
+        public UnresolvedComponent(object existingInstance)
+        {
+            var methods = existingInstance
+                .GetType()
+                .GetMethods(BindingFlags.Public | BindingFlags.Instance)
                 .Where(it => it.GetCustomAttribute(typeof(LateInitAttribute)) != null)
                 .ToArray();
 
+            if (methods.Length > 1)
+            {
+                throw new ArgumentException(
+                    $"Type {existingInstance.GetType().Name} has more than one [LateInit] annotated method. Only one method can be annotated.");
+            }
+
+            _existingInstance = existingInstance;
             if (methods.Length == 0)
             {
-                // no more missing dependencies,
-                return instance;
-            }
-            
-            foreach (var method in methods)
-            {
-                if (!ResolveMethodDependencies(method, out var resolvedDependencies))
-                {
-                    continue;
-                }
-                
-                // we have a resolved initializer
-                method.Invoke(instance, resolvedDependencies);
-                return instance;
+                _initializer = null;
+                _types = new Type[0];
+                _qualifiers = new string[0];
+                _resolvedDependencies = new object[0];
+                _isResolved = new bool[0];
+                return;
             }
 
-            return default;
+            _initializer = methods[0];
+            ReadMethodInfo(_initializer, out _types, out _qualifiers, out _resolvedDependencies, out _isResolved);
         }
 
-        private object TryConstruct(Type type, string qualifier)
+        /// <summary>
+        /// Constructor for regular components.
+        /// </summary>
+        /// <param name="type">the type of the component</param>
+        public UnresolvedComponent(Type type)
         {
-            // only public constructors are used for dependency injection
             var constructors = type.GetConstructors();
 
-            foreach (var constructor in constructors)
+            if (constructors.Length == 0)
             {
-                if (!ResolveMethodDependencies(constructor, out var resolvedDependencies))
-                {
-                    continue;
-                }
-
-                // all dependencies of this constructor have been resolved,
-                // invoke the constructor
-                return constructor.Invoke(resolvedDependencies);
+                throw new ArgumentException(
+                    $"Type {type.Name} has no public constructor. Please create a public constructor to be used for initialization.");
             }
 
-            // no constructor could be resolved, so we give up.
-            return null;
+            if (constructors.Length > 1)
+            {
+                constructors = constructors
+                    .Where(it => it.GetCustomAttribute(typeof(ConstructorAttribute)) != null)
+                    .ToArray();
+                if (constructors.Length > 1)
+                {
+                    throw new ArgumentException(
+                        $"Type {type.Name} has more than one constructor annotated with [Constructor]. Only one constructor can be annotated.");
+                }
+
+                if (constructors.Length == 0)
+                {
+                    throw new ArgumentException(
+                        $"Type {type.Name} has more than one constructor but none is annotated with [Constructor]. Please annotate exactly one with [Constructor].");
+                }
+            }
+
+
+            _initializer = constructors[0];
+            ReadMethodInfo(_initializer, out _types, out _qualifiers, out _resolvedDependencies, out _isResolved);
         }
 
-        private bool ResolveMethodDependencies(MethodBase method, out object[] dependencies)
+        /// <summary>
+        /// Helper function which reads information about a method through C# reflection and puts the results into
+        /// the given data structures.
+        /// </summary>
+        /// <param name="method">The method instance to read.</param>
+        /// <param name="types">Array containing the types of the parameters of the method.</param>
+        /// <param name="qualifiers">Array containing the qualifier of each parameter of the method.</param>
+        /// <param name="resolvedDependencies">Array with an empty space for each parameter of the method.
+        /// This is later filled when dependency resolution takes place.</param>
+        /// <param name="isResolved">Array with markers indicating if a parameter is already resolved. This will contain
+        /// one space for each parameter, all initialized to <c>false</c></param>
+        private static void ReadMethodInfo(MethodBase method,
+            out Type[] types,
+            out string[] qualifiers,
+            out object[] resolvedDependencies,
+            out bool[] isResolved)
         {
             var parameters = method.GetParameters();
-            var resolvedDependencies = new object[parameters.Length];
-            for (var index = 0; index < parameters.Length; index++)
+            var numberOfParameters = parameters.Length;
+
+            resolvedDependencies = new object[numberOfParameters];
+            types = new Type[numberOfParameters];
+            qualifiers = new string[numberOfParameters];
+            isResolved = new bool[numberOfParameters];
+
+            for (var index = 0; index < numberOfParameters; index++)
             {
                 var parameter = parameters[index];
                 var qualifierAttribute = parameter.GetCustomAttribute<QualifierAttribute>();
@@ -236,43 +504,104 @@
                     qualifier = qualifierAttribute.Name;
                 }
 
-                var parameterType = parameter.ParameterType;
-                if (_dependencies.TryGetValue(parameterType, out var dependenciesByQualifier))
-                {
-                    if (dependenciesByQualifier.TryGetValue(qualifier, out var value))
-                    {
-                        resolvedDependencies[index] = value;
-                        continue;
-                    }
-                }
-
-                // if we have no matching dependency, we have to give up.
-                dependencies = default;
-                return false;
+                qualifiers[index] = qualifier;
+                types[index] = parameter.ParameterType;
+                isResolved[index] = false;
             }
-
-            dependencies = resolvedDependencies;
-            return true;
         }
 
-        private static Type FindSuperType(IEnumerable<Type> searchSpace, Type toFind)
+        /// <summary>
+        /// Tries to resolve this component using the given resolver method.
+        /// </summary>
+        /// <param name="resolver">The resolver method to use.</param>
+        /// <param name="result">the resolved object. Only filled if resolution was successful.</param>
+        /// <returns><c>true</c> if the resolution was successful, <c>false</c> otherwise</returns>
+        /// <exception cref="InvalidOperationException">if this method is called after this component already has been
+        /// resolved.</exception>
+        public bool TryResolve(UniqueResolver resolver, out object result)
         {
-            var matches = new List<Type>();
-
-            foreach (var type in searchSpace)
+            if (_wasResolved)
             {
-                if (type == toFind)
+                throw new InvalidOperationException(
+                    "This component has already been resolved. This is almost certainly a bug.");
+            }
+
+            if (_initializer == null)
+            {
+                // we have an existing object and no [LateInit] method, so this is completely resolved.
+                result = _existingInstance;
+                _wasResolved = true;
+                return true;
+            }
+
+            var hasUnresolvedValues = false;
+
+            // walk over all dependencies and check if something can be resolved.
+            for (var i = 0; i < _resolvedDependencies.Length; i++)
+            {
+                if (_isResolved[i])
                 {
-                    return toFind;
+                    continue;
                 }
 
-                if (type.IsAssignableFrom(toFind))
+                if (resolver(_types[i], _qualifiers[i], out var value))
                 {
-                    matches.Add(type);
+                    _resolvedDependencies[i] = value;
+                    _isResolved[i] = true;
+                    continue;
+                }
+
+                hasUnresolvedValues = true;
+            }
+
+            // if all dependencies could be resolved call the constructor or [LateInit] method.
+            if (!hasUnresolvedValues)
+            {
+                if (_initializer is ConstructorInfo constructorInfo)
+                {
+                    result = constructorInfo.Invoke(_resolvedDependencies);
+                }
+                else
+                {
+                    _initializer.Invoke(_existingInstance, _resolvedDependencies);
+                    result = _existingInstance;
+                }
+
+                _wasResolved = true;
+                return true;
+            }
+
+            result = default;
+            return false;
+        }
+
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("Initializer: ");
+            sb.Append(_initializer);
+            sb.Append("\nUnresolved parameters (no match or no unique match):\n");
+
+            for (var i = 0; i < _isResolved.Length; i++)
+            {
+                if (_isResolved[i])
+                {
+                    continue;
+                }
+
+                sb.Append(" - Parameter ")
+                    .Append(i)
+                    .Append(": ")
+                    .Append(_types[i].Name);
+                if (_qualifiers[i].Length > 0)
+                {
+                    sb.Append(" (Qualifier: ")
+                        .Append(_qualifiers[i])
+                        .Append(")");
                 }
             }
 
-            return matches.FirstOrDefault();
+            return sb.ToString();
         }
     }
 }