+++ Runtime/LateInitAttribute.cs
@@ -3,7 +3,11 @@ namespace AncientLightStudios.Nanoject
     using System;
     using JetBrains.Annotations;
 
+    /// <summary>
+    /// 
+    /// </summary>
     [MeansImplicitUse]
+    [AttributeUsage(validOn:AttributeTargets.Method)]
     public class LateInitAttribute : Attribute
     {
     }