+++ Tests/Editor/GoldenDoor.cs
@@ -1,0 +1,7 @@
+namespace AncientLightStudios.Nanoject.Tests
+{
+    [Qualifier("goldenDoor")]
+    class GoldenDoor : Door
+    {
+    }
+}