+++ Tests/Editor/Palace.cs
@@ -1,0 +1,9 @@
+namespace AncientLightStudios.Nanoject.Tests
+{
+    public class Palace : House
+    {
+        public Palace([Qualifier("goldenDoor")] Door door) : base(door)
+        {
+        }
+    }
+}